+++ tests/librustzcash_vectors.rs
@@ -1,4 +1,4 @@
-use std::convert::TryFrom;
+use core::convert::TryFrom;
 
 #[macro_use]
 extern crate lazy_static;