+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Entries are listed in reverse chronological order.
 ## Unreleased
 
 * Migrate to `group` 0.12, `jubjub` 0.9, `pasta_curves` 0.4
+* Added support for `no-std` builds, via new (default-enabled) `std` and `alloc`
+  feature flags. Module `batch` is supported on `alloc` feature only. Module
+  `frost` is supported on `std` feature only.
 
 ## 0.2.0